app/src/main/java/com/leboncoin/youralbums/viewmodels/AlbumViewModel.kt
package com.leboncoin.youralbums.viewmodels

import androidx.lifecycle.*
import com.leboncoin.youralbums.repository.IRepository
import kotlinx.coroutines.launch
import java.io.IOException

/**
 * @param application The application that this viewmodel is attached to
 */

class AlbumViewModel(private val albumsRepository: IRepository) : ViewModel() {
    /**
     * Albums displayed on the screen.
     */
    val albums = albumsRepository.albums

    /**
     * Event triggered for network error. This is private to avoid exposing a
     * way to set this value to observers.
     */
    private var _eventNetworkError = MutableLiveData(false)

    /**
     * Event triggered for network error. Views should use this to get access
     * to the data.
     */
    val eventNetworkError: LiveData<Boolean>
        get() = _eventNetworkError

    /**
     * Flag to display the error message. This is private to avoid exposing a
     * way to set this value to observers.
     */
    private var _isNetworkErrorShown = MutableLiveData(false)

    /**
     * Flag to display the error message. Views should use this to get access
     * to the data.
     */
    val isNetworkErrorShown: LiveData<Boolean>
        get() = _isNetworkErrorShown

    /**
     * init{} is called immediately when this ViewModel is created.
     */
    init {
        refreshDataFromRepository()
    }

    /**
     * Refresh data from the repository. Use a coroutine launch to run in a
     * background thread.
     */
    private fun refreshDataFromRepository() {
        viewModelScope.launch {
            try {
                albumsRepository.refreshAlbums()
                _eventNetworkError.value = false
                _isNetworkErrorShown.value = false

            } catch (networkError: IOException) {
                // Show a Toast error message and hide the progress bar.
                if (albums.value.isNullOrEmpty())
                    _eventNetworkError.value = true
            }
        }
    }


    /**
     * Resets the network error flag.
     */
    fun onNetworkErrorShown() {
        _isNetworkErrorShown.value = true
    }
}

@Suppress("UNCHECKED_CAST")
class ViewModelFactory(private val repository: IRepository)
    : ViewModelProvider.NewInstanceFactory() {
    override fun <T : ViewModel> create(modelClass: Class<T>) =
        (AlbumViewModel(repository) as T)
}
